Remove outdated files
This commit deletes a number of outdated files including the global style sheet, SVG icons and JS modules that are no longer necessary for the application. This clean-up will help streamline the codebase. In the package-lock file, a package 'sass' is newly introduced, though a lot of packages are removed.