Update component styles and add features
This commit introduces updates to the Reinsurance, PartnersSlide, BurgerMenu and AdCard components' padding, margin, and other CSS properties. It introduces new components FundingBanner and MainCollection, along with their CSS files. It also includes changes to the way the most recent ads and MainCollection are generated. Updates are made to change the route for motos and the setup for a new collection page. Additionally, a new API to get vehicle data and the logic to use it are included, making the data handling more dynamic.